+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added scenario option `waitUntil`.
+- Added option `waitUntil` for action `clickWithRedirect`.
 
 ## 0.3.3 - 2023-06-21
 ### Fixed